Most application setup work completed, started work on surface extraction. Need to nail down the proper coordinate mapping method to ensure properly sampling higher dimensional spaces

+++ ComputeApplication/voxel.hpp
@@ -1,8 +1,10 @@
 #pragma once
+#include <cstdint>
+
 // Voxel Memory Structure
 // Could be more complicated, i.e. materials and blending, data packing
 // Demo just needs density for meshing
 struct Voxel
 {
-  float density;
+  uint16_t density;
 };
